chore(dx): two-tier git hooks + pnpm verify + 10x faster validate-content (pr 8 of audit)
Why
---
The audit (Theme 10 + Standard 9) flagged the per-commit DX tax:
.husky/pre-commit ran `check + typecheck + validate-content + test`
on every commit at 25-40s. Compound that across 5 incremental
commits to fix one CSS property and a developer waits 2-3 minutes
at the hook gate. The friction-to-value ratio was worst for the
typecheck + tests parts — they catch errors that CI also catches
within 2 minutes, so running them locally on every commit was pure
overhead.

Separately, the validate-content runner spawned 17 sequential `tsx`
child processes — ~200ms Node startup × 17 = ~3.4s of pure overhead
before any actual validation work. Measured baseline confirmed.

Changes
-------
.husky/pre-commit:
- Now runs ONLY `pnpm check` (biome, ~0.26s) + the copilot-sync
  conditional (sub-second when source files haven't changed).
- Per-commit tax: ~25-40s → ~0.26s.
- The copilot-sync stays here because the regenerated files must
  land in the same commit as the source change — moving it to
  pre-push would let CLAUDE.md and .github/* diverge between
  intermediate commits on a stack.

.husky/pre-push:
- Existing branch-name guard kept (fail-fast — no point running
  the full chain on a malformed branch).
- New: runs `pnpm verify` after the branch-name check. Same gates
  the prior pre-commit ran, but ONCE per push instead of once per
  commit.

package.json:
- New `verify` script: `check && typecheck && validate-content &&
  check:client-naming && test`. Pre-pr command callable directly
  from CLI.
- `ci` and `ci:local` rewritten to compose `verify` — no
  duplication of the gate list across three scripts.
- `validate-content` script: now `tsx scripts/validate-content.ts`
  (was `node scripts/validate-content.mjs`).

scripts/validate-content.ts (new) + scripts/validate-content.mjs
(deleted):
- Single tsx process iterates `await import()` over the 18 content
  files in a try/catch loop. Per-file pass/fail reporting
  preserved; the first 3 lines of any parse error are printed.
- Measured wall-clock: 3.4s+ → 0.36s (~10x faster).
- No more sequential subprocess sprawl; a single Node startup
  amortizes across all 18 files.

DECISIONS.md:
- New "Two-tier git hooks + pnpm verify + 10x faster
  validate-content" section with four ADR bullets.

Verification
------------
- pnpm verify: 6.5s (full chain — same gates as old pre-commit)
- pnpm check (new pre-commit): 0.26s
- pnpm test: 256/256 pass
- pnpm typecheck: clean
- 0 client-naming violations (strict mode)
- pnpm validate-content: 18 passed, 0 failed, 0.36s

Refs docs/audit/2026-05-19-principal-audit.md (Theme 10, Standard 9).
