+++ DECISIONS.md
@@ -99,6 +99,13 @@ ADR-lite running log. One bullet per decision · date · reversibility note.
 - **2026-05-19** · **dom-size score 0.9 not achievable for this composition.** The portfolio has 18 content-dense sections totalling ~1277 nodes post-fix. Lighthouse's "good" threshold is <800 nodes. Reaching 0.9 would require removing approximately half the content sections — defeating the purpose of the portfolio as a hiring artifact. The correct framing: the dom-size metric is calibrated for typical SPAs with 1-2 views; a 18-section portfolio is an outlier the metric doesn't model well. The `maxNumericValue: 1500` gate catches actual regressions (someone accidentally doubling DOM) while acknowledging the structural reality. _Re-tighten to error only if architectural changes (e.g., section-by-section lazy loading via route splitting) bring node count below 800. Do not re-tighten to score >= 0.9 without content reduction._
 - **2026-05-19** · **unused-javascript root cause and gate calibration.** Measured: 2 items, 43KB total wasted bytes. Both items are Next.js App Router `rootMainFiles` — framework bootstrap chunks that are unconditionally loaded before any user code: (1) `08i` chunk: ~70KB compressed, contains getDeploymentId, getAssetPrefix, InvariantError, and other Next.js initialization code — 31% unused on initial paint; (2) `0fjplx8p` chunk: ~37KB compressed, contains the App Router RSC client handler, Server Actions dispatcher, and client-side navigation runtime — 57% unused on initial paint. These chunks are required for RSC hydration and the full App Router runtime even though most of their code paths don't execute on the initial page load sequence (they're needed for navigation, form interactions, revalidation). There is no actionable fix: they cannot be dynamically imported (they run before any module load), and removing them requires abandoning App Router. The wasted-bytes figure is structurally stable — it reflects Next.js's framework bootstrap split, not application-level sloppiness. Gate set to `["warn", { "maxNumericValue": 60 }]`: the audit's `numericValue` is the overallSavingsMs estimate (measured at 40ms currently); budget of 60ms gives ~50% headroom before the gate fires, catching if framework bootstrap cost significantly regresses. _Do not re-tighten to `error` unless Next.js provides a mechanism to tree-shake unused App Router runtime paths. Revisit if Next.js ships a "partial hydration" mode that eliminates the navigation runtime on static pages._
 
+## 2026-05-19 — Two-tier git hooks + `pnpm verify` + 10× faster validate-content (PR 8 of audit roadmap)
+
+- **2026-05-19** · **`.husky/pre-commit` collapsed to `pnpm check` + copilot-sync conditional.** Previously ran `pnpm check && pnpm typecheck && pnpm validate-content && pnpm test` — measured at 25-40s per commit. Audit Standard 9 ("DX is measured in seconds per commit") flagged this as the single biggest DX friction in the harness. The heavy gates (typecheck, validate-content, check:client-naming, tests) move to `pre-push` where they fire ONCE per push instead of once per commit. Same total safety net at a fraction of the per-commit tax. Measured per-commit cost now: **~0.26s** (biome check + the conditional copilot regen, which fires only on CLAUDE.md / copilot-port config changes). The pre-push runs `pnpm verify` which measures ~6.5s for the full chain. _Reversible: revert this commit; per-commit tax returns._
+- **2026-05-19** · **`pnpm verify` added as the named pre-PR command.** `verify = check && typecheck && validate-content && check:client-naming && test`. Same chain pre-push runs, but callable directly from CLI without triggering the hook. Useful for "I want to confirm green before pushing" workflows. `ci` and `ci:local` rewritten to compose `verify` (no duplication of the gate list across three scripts). _Reversible: drop the script from package.json._
+- **2026-05-19** · **`scripts/validate-content.mjs` rewritten as `scripts/validate-content.ts` and run via `tsx`.** Old: spawned 17 sequential `tsx` child processes (~200ms Node startup × 17 = ~3.4s pure overhead) — one per content file. New: single tsx process iterates `await import()` in a loop with per-file try/catch. Measured wall-clock: 3.4s+ → **0.36s** (~10× faster). Per-file pass/fail reporting preserved. Same error message surface (first 3 lines of the parse error per failure). _Reversible: restore the .mjs from git history._
+- **2026-05-19** · **`.husky/pre-push` extended to run `pnpm verify`** AFTER the branch-name guard. Branch-name check fails fast (no point running 6.5s of tests if the branch name is wrong). The verify chain then catches typecheck/validate-content/client-naming/test regressions before they hit CI. _Reversible: comment out the `pnpm verify` line._
+
 ## 2026-05-19 — `.client.*` bulk rename + strict gate (PR 6b of audit roadmap)
 
 - **2026-05-19** · **9 `'use client'` files renamed to `*.client.{tsx,ts}`** as part of the audit Theme 4 + Standard 2 enforcement. `git mv` preserves history. The files: `components/responsive/{CRTOverlay,DesktopTopbar,Dock,MatrixRain,MobileTitleBar,StatusBar}.tsx`, `components/sections/Footer.tsx`, `lib/error-bridge.ts`, `lib/use-breakpoint.tsx`. ~15 import references updated across 8 distinct files (AppShell.client.tsx, FooterLazy.client.tsx, Footer.client.tsx, ErrorBoundary.client.tsx, InteractiveShell.tsx, app/page.tsx, plus 2 source-grep tests). The renames are mechanical but make the RSC boundary visible from the file tree alone (audit's "RSC drift visible in code review" property). _Reversible at the cost of 9 `git mv` reverts + import path updates._
