+++ package.json
@@ -21,11 +21,12 @@
     "test:e2e": "playwright test",
     "lhci": "lhci autorun",
     "lhci:mobile": "lhci autorun --config=lighthouserc.mobile.json",
-    "validate-content": "node scripts/validate-content.mjs",
+    "validate-content": "tsx scripts/validate-content.ts",
     "bundle-check": "node scripts/check-bundle-size.mjs",
     "check:client-naming": "node scripts/check-client-naming.mjs",
-    "ci": "pnpm check && pnpm typecheck && pnpm validate-content && pnpm check:client-naming && pnpm test && pnpm build && pnpm bundle-check",
-    "ci:local": "pnpm check && pnpm typecheck && pnpm validate-content && pnpm check:client-naming && pnpm test",
+    "verify": "pnpm check && pnpm typecheck && pnpm validate-content && pnpm check:client-naming && pnpm test",
+    "ci": "pnpm verify && pnpm build && pnpm bundle-check",
+    "ci:local": "pnpm verify",
     "commit": "cz",
     "prepare": "husky || true",
     "postinstall": "node scripts/strip-next-polyfills.mjs"
